requirements.txt
torch>=2.3.0
torchvision>=0.18.0
albumentations>=1.4.0
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
tqdm>=4.66.0
rich>=13.7.0
pandas>=2.2.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
flask>=3.0.0
